requirements.txt
django>=1.11.15,<2
django-report-builder
ipdb
django-rest-framework
django-rest-swagger
django-filter
djcacheutils
